Added options 'Voucher' and 'Cheque' to the payment types.
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@11387 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/src/org/mifos/application/accounts/loan/business/LoanBO.java
@@ -1651,8 +1651,8 @@ public class LoanBO extends AccountBO {
 
 			addAccountActionDate(loanScheduleEntity);
 			for (FeeInstallment feeInstallment : feeInstallmentList) {
-				if (feeInstallment.getInstallmentId() == installmentDate
-						.getInstallmentId()
+				if (feeInstallment.getInstallmentId().shortValue() == installmentDate
+						.getInstallmentId().shortValue()
 						&& !feeInstallment.getAccountFeesEntity().getFees()
 								.isTimeOfDisbursement()) {
 					LoanFeeScheduleEntity loanFeeScheduleEntity = new LoanFeeScheduleEntity(
